// projects/scion/workbench/src/lib/perspective/migration/model/workbench-perspective-migration-v3.model.ts
/*
 * Copyright (c) 2018-2024 Swiss Federal Railways
 *
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 */

export interface MPerspectiveLayoutV3 {
  referenceLayout: {
    workbenchGrid: string;
    outlets: string;
    desktop: string;
  };
  userLayout: {
    workbenchGrid: string;
    outlets: string;
    desktop: string;
  };
}